refactor:  altera 'Validatable' e 'ValidationHandler' de pacotes
- Move 'ValidationHandler' para pacote 'org.osnormais.storage.api.domain.validation.handler'
- Move 'Validatable' para pacote 'org.osnormais.storage.api.domain.validation'
