+++ domain/src/main/java/org/osnormais/storage/api/domain/Identifier.java
@@ -3,7 +3,7 @@ package org.osnormais.storage.api.domain;
 import java.util.Objects;
 
 import org.osnormais.storage.api.domain.validation.ValidationError;
-import org.osnormais.storage.api.domain.validation.ValidationHandler;
+import org.osnormais.storage.api.domain.validation.handler.ValidationHandler;
 
 public abstract class Identifier<T> implements ValueObject {
