--- domain/src/main/java/org/osnormais/storage/api/domain/Validatable.java
@@ -1,10 +1,0 @@
-package org.osnormais.storage.api.domain;
-
-import org.osnormais.storage.api.domain.validation.ValidationHandler;
-
-@FunctionalInterface
-public interface Validatable {
-
-    void validate(ValidationHandler handler);
-
-}
