+++ domain/src/main/java/org/osnormais/storage/api/domain/validation/Validatable.java
@@ -1,0 +1,10 @@
+package org.osnormais.storage.api.domain.validation;
+
+import org.osnormais.storage.api.domain.validation.handler.ValidationHandler;
+
+@FunctionalInterface
+public interface Validatable {
+
+    void validate(ValidationHandler handler);
+
+}
