+++ domain/src/main/java/org/osnormais/storage/api/domain/validation/Validator.java
@@ -1,5 +1,7 @@
 package org.osnormais.storage.api.domain.validation;
 
+import org.osnormais.storage.api.domain.validation.handler.ValidationHandler;
+
 public abstract class Validator {
 
     private final ValidationHandler handler;
